# docs/swagger.yaml
definitions:
...
  forms.ContactInfo:
    properties:
      city:
        type: string
      email:
        type: string
      phone:
        type: string
    type: object
  forms.ErrorForm:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  forms.ProfileForm:
    properties:
      contact_info:
        $ref: '#/definitions/forms.ContactInfo'
      profile:
        $ref: '#/definitions/forms.ProfileInfo'
      school:
        $ref: '#/definitions/forms.SchoolEducationForm'
      university:
        $ref: '#/definitions/forms.UniversityEducationForm'
    type: object
  forms.ProfileInfo:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      cover_url:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      sex:
        type: integer
      username:
        type: string
    type: object
...
  forms.SchoolEducationForm:
    properties:
      school_city:
        type: string
      school_name:
        type: string
    type: object
...
  forms.UniversityEducationForm:
    properties:
      faculty:
        type: string
      grad_year:
        type: integer
      univ_city:
        type: string
      univ_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/profile:
    post:
      consumes:
      - application/json
      description: Update user profile by id
      parameters:
      - description: First name
        in: formData
        name: firstname
        required: true
        type: string
      - description: Last name
        in: formData
        name: lastname
        required: true
        type: string
      - description: Birth date
        in: formData
        name: birth_date
        required: true
        type: string
      - description: Sex
        in: formData
        name: sex
        required: true
        type: integer
      - description: Bio
        in: formData
        name: bio
        required: true
        type: string
      - description: Avatar
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated
          schema:
            type: string
        "400":
          description: Failed to parse form
          schema:
            $ref: '#/definitions/forms.ErrorForm'
        "500":
          description: Failed to update profile
          schema:
            $ref: '#/definitions/forms.ErrorForm'
      summary: Update user profile
      tags:
      - Profile
  /api/profile/{id}:
    get:
      consumes:
      - application/json
      description: Get user profile by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/forms.ProfileForm'
        "400":
          description: Failed to parse user id
          schema:
            $ref: '#/definitions/forms.ErrorForm'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/forms.ErrorForm'
        "500":
          description: Failed to get profile
          schema:
            $ref: '#/definitions/forms.ErrorForm'
      summary: Get user profile
      tags:
      - Profile
...
swagger: "2.0"